Refactor Sequence into Bundle
